requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
pydantic==2.10.4
httpx==0.28.1
pyyaml==6.0.2
pywinpty==2.0.14; sys_platform == "win32"
